backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pydantic
python-dotenv
pydantic[email]